pyproject.toml
[project]
name = "repensar-multiplatform-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "sqlmodel>=0.0.24",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.1.3",
    "python-multipart>=0.0.6",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic-settings>=2.0.0",
    "pytest>=8.4.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
    "pytest-mock>=3.11.0",
]
